adding the refresh option for variants.
