+++ go.mod
@@ -3,7 +3,6 @@ module github.com/moore0n/hlstail
 go 1.13.7
 
 require (
-	github.com/ingest/manifest v0.0.0-20190812004131-f98280a4b8c1
 	github.com/urfave/cli v1.22.2
 	golang.org/x/crypto v0.0.0-20200207205829-a95e85b341fd
 )
